+++ .github/workflows/go-ci.yml
@@ -14,6 +14,7 @@ concurrency:
 
 jobs:
   setup:
+    name: Setup
     runs-on: ubuntu-latest
     outputs:
       debug: ${{ steps.debug.outputs.debug }}
@@ -104,11 +105,10 @@ jobs:
           file: ./coverage.txt
           name: coverage-${{ matrix.os }}
 
-  unit_race_test:
+  unit_test_race:
     needs: [lint, go_mod_tidy_check]
-    name: Run Unit Tests with Race Detector
+    name: Unit Tests with Race Detector (ubuntu-latest)
     runs-on: ubuntu-latest
-    continue-on-error: true
 
     steps:
       - uses: actions/checkout@v4
@@ -119,24 +119,11 @@ jobs:
           go-version: ${{ inputs.go-version }}
 
       - name: execute test run
-        run: make test-unit-race
+        run: make test-unit-race ENABLE_VERBOSE=${{ needs.setup.outputs.debug }}
 
   integration_test:
+    name: Integration Tests
     needs: [lint, go_mod_tidy_check]
-    name: Run Integration Tests
-    runs-on: ubuntu-latest
-    continue-on-error: true
-
-    steps:
-      - uses: actions/checkout@v4
-
-      - name: set up go
-        uses: actions/setup-go@v5
-        with:
-          go-version: ${{ inputs.go-version }}
-
-      - name: Integration Tests
-        run: make test-integration
-
-      - name: Integration Tests with Race Detector
-        run: make test-integration-race
+    uses: ./.github/workflows/integration-tests.yml
+    with:
+      go-version: ${{ inputs.go-version }}
