+++ Makefile
@@ -3,6 +3,8 @@ PROJECTNAME=$(shell basename "$(PWD)")
 DIR_FULLPATH=$(shell pwd)
 versioningPath := "github.com/celestiaorg/celestia-node/nodebuilder/node"
 LDFLAGS=-ldflags="-X '$(versioningPath).buildTime=$(shell date)' -X '$(versioningPath).lastCommit=$(shell git rev-parse HEAD)' -X '$(versioningPath).semanticVersion=$(shell git describe --tags --dirty=-dev 2>/dev/null || git rev-parse --abbrev-ref HEAD)'"
+TAGS=integration
+SHORT=
 ifeq (${PREFIX},)
 	PREFIX := /usr/local
 endif
@@ -12,6 +14,11 @@ ifeq ($(ENABLE_VERBOSE),true)
 else
 	VERBOSE =
 	LOG_AND_FILTER =
+endif
+ifeq ($(SHORT),true)
+	INTEGRATION_RUN_LENGTH = -short
+else
+	INTEGRATION_RUN_LENGTH =
 endif
 ## help: Get more info on make commands.
 help: Makefile
@@ -118,28 +125,20 @@ test-unit:
 ## test-unit-race: Running unit tests with data race detector
 test-unit-race:
 	@echo "--> Running unit tests with data race detector"
-	@go test -race `go list ./... | grep -v nodebuilder/tests`
+	@go test $(VERBOSE) -race -covermode=atomic -coverprofile=coverage.txt `go list ./... | grep -v nodebuilder/tests` $(LOG_AND_FILTER)
 .PHONY: test-unit-race
 
 ## test-integration: Running /integration tests located in nodebuilder/tests
 test-integration:
-	@echo "--> Running integrations tests"
-	@go test ./nodebuilder/tests
+	@echo "--> Running integrations tests $(VERBOSE) -tags=$(TAGS) $(INTEGRATION_RUN_LENGTH)"
+	@go test $(VERBOSE) -tags=$(TAGS) $(INTEGRATION_RUN_LENGTH) ./nodebuilder/tests
 .PHONY: test-integration
 
 ## test-integration-race: Running integration tests with data race detector located in node/tests
 test-integration-race:
-	@echo "--> Running integration tests with data race detector"
-	@go test -race ./nodebuilder/tests
+	@echo "--> Running integration tests with data race detector -tags=$(TAGS)"
+	@go test -race -tags=$(TAGS) ./nodebuilder/tests
 .PHONY: test-integration-race
-
-## test: Running both unit and integrations tests
-test:
-	@echo "--> Running all tests without data race detector"
-	@go test ./...
-	@echo "--> Running all tests with data race detector"
-	@go test -race ./...
-.PHONY: test
 
 ## benchmark: Running all benchmarks
 benchmark:
@@ -163,7 +162,6 @@ pb-gen:
 		done; \
 	done;
 .PHONY: pb-gen
-
 
 ## openrpc-gen: Generate OpenRPC spec for Celestia-Node's RPC api
 openrpc-gen:
@@ -221,7 +219,6 @@ goreleaser-release:
 .PHONY: goreleaser-release
 
 # Copied from https://github.com/dgraph-io/badger/blob/main/Makefile
-
 USER_ID      = $(shell id -u)
 HAS_JEMALLOC = $(shell test -f /usr/local/lib/libjemalloc.a && echo "jemalloc")
 JEMALLOC_URL = "https://github.com/jemalloc/jemalloc/releases/download/5.2.1/jemalloc-5.2.1.tar.bz2"
