+++ nodebuilder/tests/nd_test.go
@@ -1,3 +1,5 @@
+//go:build nd || integration
+
 package tests
 
 import (
