+++ share/p2p/discovery/discovery_test.go
@@ -1,3 +1,5 @@
+// go:build !race
+
 package discovery
 
 import (
