+++ state/core_access_test.go
@@ -1,3 +1,5 @@
+//go:build !race
+
 package state
 
 import (
